pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = ""
authors = [
    {name = "McEazy2700",email = "codeepoch@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-stubs (>=5.1.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-yasg[validation] (>=1.21.10,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
reportUnknownVariableType = false
reportUnannotatedClassAttribute = false
reportImplicitOverride = false
reportUnknownMemberType = false
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnsafeMultipleInheritance = false
reportDeprecated = false
reportAny = false
reportIncompatibleMethodOverride = false
reportExplicitAny = false
reportUnnecessaryComparison = false
reportIncompatibleVariableOverride = false
